## Description

Fix #3511

Improved grover runtime.

Summary of changes:
- updated grover featurizer to pre-compute grover graph features during dataset featurization phase
- added `BatchGroverGraph` to handle edge cases in batching of grover molecular graph
- minor changes to grover torch model

Testing setup:

Accelerator: Apple M1 GPU
Dataset: Delaney (1127 data points)
Other parameters: Batch size 100, number of epochs: 10

Testing results:

 | Approach | Total training time | Avg. time for batch preperation | Avg. time for extracting attributes | Avg. time on forward pass | Avg. time for an iteration |
 | --- | ---- | ---- | ---- | ---- | ---- |
 | New | 454 sec | 0.007 sec | 2.68e-06 sec | 1.043 sec | 1.05 sec |
 | Old | 138 sec | 0.002 sec | 2.65 sec | 1.026 sec | 3.682 sec

Total training time on old approach: 454 seconds
Total training time on new approach: 138 seconds (~3.2x speedup)

Average time corresponds to time taken for an iteration.

## Type of change

Please check the option that is related to your PR.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
  - In this case, we recommend to discuss your modification on GitHub issues before creating the PR
- [ ] Documentations (modification for documents)

## Checklist

- [ ] My code follows [the style guidelines of this project](https://deepchem.readthedocs.io/en/latest/development_guide/coding.html)
  - [x] Run `yapf -i <modified file>` and check no errors (**yapf version must be  0.32.0**)
  - [ ] Run `mypy -p deepchem` and check no errors
  - [x] Run `flake8 <modified file> --count` and check no errors
  - [ ] Run `python -m doctest <modified file>` and check no errors
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New unit tests pass locally with my changes
- [ ] I have checked my code and corrected any misspellings